# backend/app/schemas/problem.py
from datetime import datetime

from pydantic import BaseModel, Field, model_validator

from app.schemas.case_training import DIMENSION_SPECS, CaseDefinition, CaseOpening, CaseRubric
from app.schemas.conversation import MessageIn, MessageRead


class ProblemCreate(BaseModel):
    type: str = Field(min_length=1, max_length=40)
    title: str = Field(min_length=1, max_length=200)
    description: str = Field(default="", max_length=4000)
    target: str = Field(default="all", pattern="^(all|class|individual)$")
    target_label: str = Field(default="全体学生", max_length=200)
    target_ids: list[str] = Field(default_factory=list, max_length=200)
    content_type: str = Field(default="question", pattern="^(question|guided_case)$")
    slug: str | None = Field(default=None, min_length=1, max_length=120)
    specialty: str = Field(default="", max_length=80)
    difficulty: str = Field(default="basic", max_length=30)
    estimated_minutes: int = Field(default=10, ge=1, le=180)
    version: int = Field(default=1, ge=1)
    parent_problem_id: int | None = Field(default=None, gt=0)
    case_definition: CaseDefinition | None = None
    rubric: CaseRubric | None = None
    capability_tags: list[str] = Field(default_factory=list, max_length=6)

    @model_validator(mode="after")
    def validate_case(self) -> "ProblemCreate":
        if self.content_type == "guided_case" and (self.case_definition is None or self.rubric is None):
            raise ValueError("guided_case requires case_definition and rubric")
        if self.content_type == "guided_case" and not self.slug:
            raise ValueError("guided_case requires slug")
        if self.content_type == "question" and (self.case_definition is not None or self.rubric is not None):
            raise ValueError("question cannot include case fields")
        if any(tag not in {item[0] for item in DIMENSION_SPECS} for tag in self.capability_tags):
            raise ValueError("invalid capability tag")
        return self


class ProblemUpdate(ProblemCreate):
    status: str | None = Field(default=None, pattern="^(draft|published|rejected)$")

    @model_validator(mode="after")
    def reject_guided_status_override(self) -> "ProblemUpdate":
        if self.content_type == "guided_case" and self.status is not None:
            raise ValueError("guided_case status is controlled by review and publish endpoints")
        return self


class ProblemRead(BaseModel):
    id: int
    type: str
    title: str
    description: str = ""
    target: str
    target_label: str
    target_ids: list[str] = Field(default_factory=list)
    status: str
    created_at: datetime
    published_at: datetime | None = None
    answer_count: int = 0
    content_type: str = "question"
    slug: str | None = None
    specialty: str = ""
    difficulty: str = "basic"
    estimated_minutes: int = 10
    version: int = 1
    parent_problem_id: int | None = None
    author_id: int | None = None
    medical_review_status: str = "not_submitted"
    opening: CaseOpening | None = None
    capability_tags: list[str] = Field(default_factory=list)


class ProblemAuthoringRead(ProblemRead):
    case_definition: CaseDefinition | None = None
    rubric: CaseRubric | None = None


class QuestionThreadUpsert(BaseModel):
    messages: list[MessageIn] = Field(default_factory=list, max_length=100)


class QuestionThreadRead(BaseModel):
    question_id: int
    messages: list[MessageRead] = []
    updated_at: datetime


class StudentQuestionRead(BaseModel):
    id: int
    type: str
    title: str
    description: str = ""
    published_at: datetime
    status: str = Field(pattern="^(answered|unanswered)$")


class StudentQuestionPage(BaseModel):
    items: list[StudentQuestionRead]
    total: int
    limit: int
    offset: int